+++ packages/connectors/src/http-error.ts
@@ -1,0 +1,9 @@
+export class HttpError extends Error {
+  public readonly code: string;
+
+  constructor(code: string, message: string) {
+    super(message);
+    this.name = "HttpError";
+    this.code = code;
+  }
+}